backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.34.0
botocore==1.34.162
celery==5.3.4
certifi==2025.4.26
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
distro==1.9.0
Django==5.0.1
django-celery-beat==2.7.0 
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.5
django-ratelimit==4.1.0
django-redis==5.4.0
django-storages==1.14.2
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dj-database-url==2.1.0
drf-yasg==1.21.7
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx>=0.28.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kombu==5.5.3
msgpack==1.1.0
openai>=1.50.0
packaging==25.0
Pillow==10.1.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pgbouncer-python==0.1.0
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
cryptography==43.0.3
zxcvbn==4.4.28
user-agents==2.2.0
pyotp==2.9.0
pypng==0.20220715.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
s3transfer==0.9.0
sentry-sdk==1.39.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0

# Report Generation
reportlab==4.0.4
xlsxwriter==3.1.2
stripe==7.10.0
# mercadopago==2.2.1  # Removed - Stripe-only integration
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
scipy==1.12.0